static array loop with integer size comparison is has been implemented
Traversal of loop body to find the valid array container being index
was finished.
Extracting the size of the array from intialiser list or
the array being declared with an int e.g int a[10].
Is comparitor valid is comparing the size of the array and the other
supplied value.
Checking safety of loop body and replacement of usages in body are
reused code.